refactor: 패키지별로 클래스 구분
- domain package
- view package
- checker package
