+++ src/main/java/baseball/checker/CommandChecker.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.checker;
 
 public final class CommandChecker {
